+++ apps/meteor/ee/server/services/CHANGELOG.md
@@ -1,5 +1,31 @@
 # rocketchat-services
 
+## 1.1.35
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@6.9.2
+  - @rocket.chat/rest-typings@6.9.2
+  - @rocket.chat/core-services@0.3.17
+  - @rocket.chat/model-typings@0.4.3
+  - @rocket.chat/models@0.0.41
+  </details>
+
+## 1.1.34
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@6.9.1
+  - @rocket.chat/rest-typings@6.9.1
+  - @rocket.chat/core-services@0.3.16
+  - @rocket.chat/model-typings@0.4.2
+  - @rocket.chat/models@0.0.40
+  </details>
+
 ## 1.1.33
 
 ### Patch Changes